Separerat dal för create och index för film
