+++ Pages/Films/Create.cshtml.cs
@@ -1,10 +1,8 @@
-﻿using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.EntityFrameworkCore;
-using RetroTapes.Data;
+//using Microsoft.EntityFrameworkCore;
+//using RetroTapes.Data;
 using RetroTapes.Services;
 using RetroTapes.ViewModels;
 
@@ -12,9 +10,13 @@ namespace RetroTapes.Pages.Films
 {
     public class CreateModel : PageModel
     {
-        private readonly SakilaContext _db;
+        //private readonly SakilaContext _db;
         private readonly FilmService _svc;
-        public CreateModel(SakilaContext db, FilmService svc) { _db = db; _svc = svc; }
+        private readonly LookupService _lookups;
+        //public CreateModel(SakilaContext db, FilmService svc) { _db = db; _svc = svc; }
+
+        public CreateModel(FilmService svc, LookupService lookups) { _svc = svc; _lookups = lookups; }
+
 
         [BindProperty] public FilmEditVm Vm { get; set; } = new();
 
@@ -43,16 +45,20 @@ namespace RetroTapes.Pages.Films
         private async Task PopulateDropdownsAsync()
         {
             LanguageOptions = new SelectList(
-                await _db.Languages.AsNoTracking().OrderBy(l => l.Name).ToListAsync(),
-                "LanguageId", "Name");
+                //await _db.Languages.AsNoTracking().OrderBy(l => l.Name).ToListAsync(),
+                //"LanguageId", "Name");
+                await _lookups.GetLanguagesAsync(), "LanguageId", "Name");
+
 
             CategoryOptions = new MultiSelectList(
-                await _db.Categories.AsNoTracking().OrderBy(c => c.Name).ToListAsync(),
-                "CategoryId", "Name");
+                //await _db.Categories.AsNoTracking().OrderBy(c => c.Name).ToListAsync(),
+                //"CategoryId", "Name");
+                await _lookups.GetCategoriesAsync(), "CategoryId", "Name");
 
             ActorOptions = new MultiSelectList(
-                await _db.Actors.AsNoTracking().OrderBy(a => a.LastName).ThenBy(a => a.FirstName).ToListAsync(),
-                "ActorId", "LastName");
+                //await _db.Actors.AsNoTracking().OrderBy(a => a.LastName).ThenBy(a => a.FirstName).ToListAsync(),
+                //"ActorId", "LastName");
+                await _lookups.GetActorsAsync(), "ActorId", "LastName");
 
             ViewData["LanguageOptions"] = LanguageOptions;
             ViewData["CategoryOptions"] = CategoryOptions;
