+++ Services/FilmService.cs
@@ -1,11 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 using RetroTapes.Data;
+using RetroTapes.Infrastructure;
 using RetroTapes.Models;
 using RetroTapes.ViewModels;
+
 
 namespace RetroTapes.Services
 {
@@ -75,6 +73,65 @@ namespace RetroTapes.Services
 
             await _db.SaveChangesAsync();
             return (film, created);
+        }
+
+        public async Task<PagedResult<FilmListItemVm>> SearchAsync(string q, int? languageId, int? categoryId, string? sort, int pageIndex, int pageSize)
+        {
+            var query = _db.Films.AsNoTracking().AsQueryable();
+
+            if (!string.IsNullOrWhiteSpace(q))
+            {
+                var term = q.Trim();
+                query = query.Where(f => f.Title.Contains(term)
+                    || (f.Description != null && f.Description.Contains(term)));
+            }
+
+            if (languageId.HasValue)
+                query = query.Where(f => f.LanguageId == languageId.Value);
+
+            if (categoryId.HasValue)
+                query = query.Where(f => f.FilmCategories.Any(fc => fc.CategoryId == categoryId.Value));
+
+            query = sort switch
+            {
+                "title" => query.OrderBy(f => f.Title),
+                "title_desc" => query.OrderByDescending(f => f.Title),
+                "year" => query.OrderBy(f => f.ReleaseYear),
+                "year_desc" => query.OrderByDescending(f => f.ReleaseYear),
+                _ => query.OrderByDescending(f => f.LastUpdate)
+            };
+
+            var total = await query.CountAsync();
+
+            // Clamp page så vi inte skippas förbi slutet
+            var totalPages = (int)Math.Ceiling(total / (double)pageSize);
+            if (totalPages == 0) totalPages = 1;
+            if (pageIndex < 1) pageIndex = 1;
+            if (pageIndex > totalPages) pageIndex = totalPages;
+
+            var items = await query
+                .Skip((pageIndex - 1) * pageSize)
+                .Take(pageSize)
+
+                .Select(f => new FilmListItemVm
+                {
+                    FilmId = f.FilmId,
+                    Title = f.Title,
+                    ReleaseYear = f.ReleaseYear,
+                    LanguageName = f.Language.Name,
+                    Categories = string.Join(", ", f.FilmCategories.Select(fc => fc.Category.Name).OrderBy(n => n)),
+                    LastUpdate = f.LastUpdate
+                })
+                .ToListAsync();
+
+            return new PagedResult<FilmListItemVm>
+            {
+                Items = items,
+                TotalCount = total,
+                Page = pageIndex,
+                PageSize = pageSize
+            };
+
         }
 
     }
